Refactor doc and xlsx parsing
