+++ core-aspose/src/main/scala-2/bridges/aspose/word/WordToPdfAsposeBridge.scala
@@ -6,9 +6,6 @@ import models.FileContent
 import types.MimeType
 import types.MimeType.{ApplicationMsWord, ApplicationPdf}
 
-import com.aspose.words.Document
-import com.aspose.words.SaveFormat
-
 import java.io.{ByteArrayInputStream, ByteArrayOutputStream}
 import scala.reflect.ClassTag
 
@@ -21,14 +18,4 @@ object WordToPdfAsposeBridge extends SimpleBridge[ApplicationMsWord.type, Applic
   override implicit val mTag: ClassTag[M] = implicitly[ClassTag[M]]
   implicit val iTag: ClassTag[ApplicationMsWord.type] = implicitly[ClassTag[ApplicationMsWord.type]]
   implicit val oTag: ClassTag[ApplicationPdf.type] = implicitly[ClassTag[ApplicationPdf.type]]
-  
-  /**
-   * Scala 2 specific implementation of the document conversion using Aspose
-   */
-  override protected def convertDocToPdf(inputStream: ByteArrayInputStream): ByteArrayOutputStream = {
-    val asposeDoc = new Document(inputStream)
-    val pdfOutput = new ByteArrayOutputStream()
-    asposeDoc.save(pdfOutput, SaveFormat.PDF)
-    pdfOutput
-  }
 }
