+++ core-aspose/src/main/scala-3/bridges/aspose/word/WordToPdfAsposeBridge.scala
@@ -6,22 +6,10 @@ import models.FileContent
 import types.MimeType
 import types.MimeType.{ApplicationMsWord, ApplicationPdf}
 
-import com.aspose.words.{Document as AsposeDocument, SaveFormat}
-
 import java.io.{ByteArrayInputStream, ByteArrayOutputStream}
 
 /**
  * Scala 3 implementation of WordToPdfAsposeBridge.
- * Extends the common implementation with Scala 3 specific syntax.
+ * Extends the common implementation without any Scala 3 specific customizations.
  */
-object WordToPdfAsposeBridge extends SimpleBridge[ApplicationMsWord.type, ApplicationPdf.type] with WordToPdfAsposeBridgeImpl {
-  /**
-   * Scala 3 specific implementation of the document conversion using Aspose
-   */
-  override protected def convertDocToPdf(inputStream: ByteArrayInputStream): ByteArrayOutputStream = {
-    val asposeDoc = new AsposeDocument(inputStream)
-    val pdfOutput = new ByteArrayOutputStream()
-    asposeDoc.save(pdfOutput, SaveFormat.PDF)
-    pdfOutput
-  }
-}
+object WordToPdfAsposeBridge extends SimpleBridge[ApplicationMsWord.type, ApplicationPdf.type] with WordToPdfAsposeBridgeImpl
